Create text boxes
Create text boxes that display the calculation currently being performed and its result. Every tick, if a button is pressed, an update flag is set. When the update loop is complete, if this update flag has been tripped, then the result of the calculation currently being processed is calculated and displayed (if the result is invalid "invalid" is displayed instead.